auto_analyst_pro/requirements.txt
streamlit
pandas
google-genai
matplotlib
joblib
python-dotenv
scikit-learn
reportlab
